Adds Beginning Framework for React Portion
Adds the styled-components npm package and begins to style the landing
page for the application.

Adds a GET call on componentDidMount to load the data from room ID 2
onto the page and into state.

Outlines conditional render for room costs when check-in and checkout
dates are added.

Adds Jest test to make sure that the page renders without crashing.