Adds the CheckIn/CheckOut and Calendar Components
This adds the beginning functionality for the calendar. There is now
a checkin and checkout div that once clicked will trigger the rendering
of the calendar.

The calendar now populates in a table for one month. Next is adding the
ability to change months and add a second month to the div.