Adds Additional Functionality for Date Checking
This commit adds a function that generates a Moment object based on the
date and calendar selected. Then from there it will check if it is
booked during the initial render and flag if they are booked.

Adds another function that accepts a click on the calendar and if it is
not booked it will update the styling of the calendar date and update
the checkIn or checkOut date in the App components state.