Adds POST Request to Module
Adds an axios POST request to the App component that is fired when
clicking on the 'Reserve' button. Also added a POST route in the server
that listens for the POST but does nothing.

All tests are successful locally but the server test is failing on CI,
which might be due to the fact that the DB is on my local.