Adds a Modal Component
Adds a Modal Component that uses a conditional render to appear on the
screen when the calendarToggle is clicked. Once rendered it greys out
the background and pushes the CalendarPopUp module to the forefront.

Refactors the App component to use the Modal component in the render
and gets rid of the calendarCheck function.