Adds Small Styling Changes
Adds small styling changes that address spacing or the width of certain
components. It has improved the overall feel of the page but there is
still progress to be made.

Creates a Close component to add to the calendar pop-up so that it will
allow the user to exit easily.